+++ LED_DotMatrix/env.h
@@ -10,6 +10,6 @@
 #define DATA_PIN 10
 
 void write_data_to_reg(unsigned char reg, unsigned char data);
-void DEBUG(int x, int code);
+void DEBUG(long x, int code);
 
 #endif
